+++ src/ts/constant.ts
@@ -3,8 +3,8 @@
  * 
  * A JavaScript library for converting between HTML and JSON, with binding, templating, attributes, and CSS support.
  * 
- * @file        constant.js
- * @version     v2.0.0
+ * @file        constant.ts
+ * @version     v2.1.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2024
@@ -12,5 +12,5 @@
 
 
 export namespace Constant {
-    export const JHSON_JS_ATTRIBUTE_NAME = "data-jhson-js";
+    export const JHSON_JS_ATTRIBUTE_NAME: string = "data-jhson-js";
 }