+++ src/ts/data/default.ts
@@ -1,0 +1,98 @@
+/**
+ * JHson.js
+ * 
+ * A JavaScript library for converting between HTML and JSON, with binding, templating, attributes, and CSS support.
+ * 
+ * @file        default.ts
+ * @version     v2.1.0
+ * @author      Bunoon
+ * @license     MIT License
+ * @copyright   Bunoon 2024
+ */
+
+
+import { type Configuration, type StringToJson } from "../type";
+import { Char } from "./enum";
+import { Is } from "./is";
+
+
+export namespace Default {
+    export function getAnyString( value: any, defaultValue: string ) : string {
+        return typeof value === "string" ? value : defaultValue;
+    }
+
+    export function getString( value: any, defaultValue: string ) : string {
+        return Is.definedString( value ) ? value : defaultValue;
+    }
+
+    export function getBoolean( value: any, defaultValue: boolean ) : boolean {
+        return Is.definedBoolean( value ) ? value : defaultValue;
+    }
+
+    export function getNumber( value: any, defaultValue: number ) : number {
+        return Is.definedNumber( value ) ? value : defaultValue;
+    }
+
+    export function getFunction( value: any, defaultValue: object ) : any {
+        return Is.definedFunction( value ) ? value : defaultValue;
+    }
+
+    export function getArray( value: any, defaultValue: any[] ) : any[] {
+        return Is.definedArray( value ) ? value : defaultValue;
+    }
+
+    export function getObject( value: any, defaultValue: object ) : any {
+        return Is.definedObject( value ) ? value : defaultValue;
+    }
+
+    export function getStringOrArray( value: any, defaultValue: string[] ) : string[] {
+        let result: string[] = defaultValue;
+
+        if ( Is.definedString( value ) ) {
+            const values: string[] = value.toString().split( Char.space );
+
+            if ( values.length === 0 ) {
+                value = defaultValue;
+            } else {
+                result = values;
+            }
+
+        } else {
+            result = getArray( value, defaultValue );
+        }
+
+        return result;
+    }
+
+    export function getObjectFromString( objectString: any, configuration: Configuration ) : StringToJson {
+        const result: StringToJson = {
+            parsed: true,
+            object: null
+        } as StringToJson;
+
+        try {
+            if ( Is.definedString( objectString ) ) {
+                result.object = JSON.parse( objectString );
+            }
+
+        } catch ( e1: any ) {
+            try {
+                result.object = eval( `(${objectString})` );
+
+                if ( Is.definedFunction( result.object ) ) {
+                    result.object = result.object();
+                }
+                
+            } catch ( e2: any ) {
+                if ( !configuration.safeMode ) {
+                    console.error( configuration.text!.objectErrorText!.replace( "{{error_1}}",  e1.message ).replace( "{{error_2}}",  e2.message ) );
+                    result.parsed = false;
+                }
+                
+                result.object = null;
+            }
+        }
+
+        return result;
+    }
+}